Explicitly search for the current clipping path node first.
The reason for this is because when there is no document associated with a
node, 'getNodeSet()' searches relative to that node. When there is in fact
a document associated, 'getNodeSet()' searches relative to that document.
We now have a document associated with our node, so we need to make the XPath
search explicitly below the current clipping path node.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/gridsvg@317 edb9625f-4e0d-4859-8d74-9fd3b1da38cb